pull subcomponents out of GardenPage, tweak tab interface
